Add example modifiers and other refactors
Refactors:
- Change Bullet prefab to use ConstantForce instead of gravity
- Change health to be float
- Move all modifier scripts to new modifiers folder
- Fix inheritance structure of ModifierBase
- Fix bug where empty modifiers caused errors

Features:
- Add Machinegun Modifier
- Add Shotgun Modifier
- Add Mortar Modifier
- Add more accurate targeting (credit: Forest Smith)

Todo next time: Refactor additive and multiplicative modifiers. Make it so all modifiers have 3 options: add, mult, or custom. Make it so multiplicative modifiers don't stack and instead add to each other.

+++ Assets/Scripts/Enemy Scripts/EnemyHealth.cs
@@ -4,25 +4,25 @@ public class EnemyHealth : MonoBehaviour
 {
 
     public InfoBar bar;
-    public int health = 100;
-    private int maxHealth;
+    public float health = 100;
+    private float maxHealth;
 
 
     void Start()
     {
         maxHealth = health;
-        bar.setMaxValue(maxHealth);
-        bar.setValue(maxHealth);
+        bar.setMaxValue((int)maxHealth);
+        bar.setValue((int)maxHealth);
     }
 
-    public void Damage(int damage)
+    public void Damage(float damage)
     {
         if (health - damage <= 0)
         {
             Die();
         }
         health -= damage;
-        bar.setValue(health);
+        bar.setValue((int)health);
     }
 
     void Die()
